make sure the mutex won't be locked before being inited, provide more assurances that certain things can't be executed while init is happening
now the init_done variable is called init_state and shows three states:
	INIT_DONE
	INIT_NOT_DONE
	INIT_IN_PROGRESS

the difference from before is that the distinction can now be made between
"init has not not been run at all yet" and "init is currently being run".
We can therefore more accurately check whether init has not been run yet.

add init_state checks at the beginning of all API functions to prevent the
mutex from being locked before being inited if an API that locks the mutex
is erroneously called before ltm_init is called. we can't rely on
DIE_ON_INVAL_TID, since it accesses descs and thus has to be used only
when the mutex is locked

set init_state to INIT_NOT_DONE at the VERY end of ltm_uninit, since it wouldn't
be safe for ltm_init to be allowed to start until the mutex is uninitialized