fix end-of-line wrapping behavior (allows you to use the entire line without wrapping)
previously, I was assuming that the behavior readline produces (wrapping when the last character of the line is written)
is the usual behavior. turns out it isn't - the usual behavior is to stay on the last character when it is written
(don't wrap yet). if/when another character is printed, the cursor is wrapped and the character gets printed there.