fix a rather nasty bug in the idarr code which prevented it from working at all
it's not possible to subscript the pointer when it's of type struct idarr *, because that struct
is going to be a lot smaller than the data the pointer actually points to. instead, do some rather
nasty pointer arithmetic which takes into account the real size.

it's all very elegant in concept, I swear...!