Remove the concept of "type" and replace it with {left,right}bound values, which can serve the same purpose.
By setting the bounds to the first and last columns, we can create an area. By setting the leftbound to the X
value of the start coordinate and setting the rightbound the X value of the end coordinate, we can create a
rect. This new system generalizes the two representations, and thus prevents us from having to provide two
different behaviors for rects and areas all over the place.