Instead of assuming that the region referenced by a range should be copied from somewhere else, allow more direct actions to be specified.
This patch adds "action" and "val" members to struct range. action can have the values ACT_{COPY,CLEAR,SCROLL}, and val is used to store
extra information about the action. It is currently only to be used with ACT_SCROLL to specify the number of lines to scroll. Only
ACT_COPY is used currently.