Introduce an entirely new way of managing screens. This new method allows the dynamic creation of arbitrary screens, rather than only two pre-assigned ones.
This does NOT mean any change for the application using libterm, but the internal API has changed.
In terms of data structures, quite a few of the properties stored in struct ltm_term_desc have now
been moved to a new "struct screen", such as the cursor position and the contents of the screen
matrix. Instead of having a variable specifying the current screen with a couple of #defines, there
are now two types of "current" screen. There is simply the current screen, which is the screen which
is exposed to the application using libterm. There is also the current input screen, which is where
user input and output of the programs running inside the terminal goes. These variables store flexible
"screen IDs" (sid), which are indeces into an array of struct screens. There are also convenience macros
provided for concise referencing of an arbitrary screen, the current screen, or the current input
screen for a tid. The bulk of this commit consists of changing existing code to use these macros.

In terms of new functions, the main ones is "screen_alloc", which allocates a new screen and
returns its sid. There is also "screen_make_current", which sets the current screen,
"screen_give_input_focus", which sets the current input screen, and "screen_set_autoscroll",
which changes the autoscroll setting for a screen (autoscroll specifies whether a screen should be
scrolled when its cursor tries to go below the botton of the screen). In addition, many functions
(including all of the cursor functions) were modified to take an extra "sid" parameter.

Note that there is at least one major problem with this code: the screen_make_current function
currently does not have any way of instructing the application using libterm to read-load the
entire screen once the current screen has been changed.