Add a mechanism for keeping track of changes to screens that are not the main window screen
The main changes in this commit are the addition of a "scr_ups" member to the terminal
description structure, and the addition of a "record_update" function.

"scr_ups" is an array which stores information about the updates that happened to
all of the screens which were input screens during output processing.

The job of the "record_update" function is to figure out where the correct place is to
record an update given which screen is being updated. If the screen being updated is the
window screen, pretty much the same thing is done as was done previously. If the screen
being updated is the current input screen, information about the update is written to scr_ups.
record_update can also update cursor and scrolling information.

The following functions have been changed to use record_update:
	cursor_visibility
	cursor_abs_move
	screen_set_point
	screen_scroll

Note that, currently, the information stored in scr_ups is not used for anything.