Add functions for doing common things to certain ranges in screens
This commit adds four functions:

screen_inject_update:
This serves a similar purpose to screen_set_point, but it allows
adding an entire range at one time, instead of doing it point-by-
point. It is currently only used by screen_scroll_rect (described
below).

screen_clear_range:
This is for clearing (setting to all spaces) a region of a screen.

screen_copy_range:
This is for copying a region of a screen to an identically-shaped
region of another screen.

screen_scroll_rect:
This for scrolling (moving all lines up) a rectangle-shaped region
of a screen.

This commit also adds a "struct link" structure, which defines how
screens relate to each other. It is currently used in the
screen_copy_range function, but will be used for other things in the
future (it's a secret!).

Note that the last three functions are not actually used by anything
yet, but they will be in the future.