Add a machanism for having the contents of a screen appear on another screen (aka "linking")
This commit adds the possibility for terminal emulation modules and other things that
have access to libterm's internal API (i.e. not applications that use it to do terminal
emulation) to "link" screens together in such a way that one screen will appear on another
one, and all updates that happen to the screen being linked will be copied to the screen
being linked to. The immediate reason that I wrote this code is because of a certain
feature of the IBM3270 that I thought would probably require this to be implemented.
I imaging this feature will have other uses, however. This code does not effect existing
paths very much, and therefore will not effect the speed of the more typical no links
case.

The API used to access this functionality is quite simple, it involves only two functions:
screen_{add,del}_link. This linking information is stores in two new members of struct
screen: "downlinks" (which screens link to this screen), and "uplinks" (which screens this
screen links to). {down,up}links are both linked lists which store struct links which
describe the links.

The dirty details of how this is implemented are in the additions to io.c. Basically
when output processing is done, the updates stored in scr_ups are looked at, and the
updates are copied to all of the screens that each of the updated screens links to. The
dirty details are available for all to see in propagate_updates and translate_ranges.

One thing that I'd like this to be able to do is to detect circular links. I already have
quite a bit of code in screen_add_link to check whether the link makes any sense, but
it is not capable of detecting when someone tries to make a circular link. I'd like it
to be able to do this because circular links would probably cause infinite recursion or
something nasty like that. I'd also like it to be easier to move link. This would probably
involve having the remove link function automatically clear the stuff that had previously
been copied.

On a slightly related note, this code has remained uncommitted for far too long. While I
semi-recently got around to committing some of the work which stands on its own, the
actual linking stuff has been sitting in my working directory for some time. In the future,
I'd like to adopt the "commit early, commit often" strategy. This will ensure that all
of the changes I make are recorded, and that large chunks of code don't build up in my
working directory.

This fixes #11.