Add the infrastructure necessary for screen_make_current to do something useful.
How this works:

During output processing, the screen which libterm considers to be
"current" does not change. It is stored in descs[tid].old_cur_screen (a
new member of struct term_desc), and is set to descs[tid].cur_screen
near the beginning of ltm_process_output. Then, once output processing
is finished, libterm checks whether old_cur_screen and cur_screen are
the same. If they are, than the updates stored for old_cur_screen can be
sent to the program. If not, there's no way of knowing what changed
because we're on a completely different screen now. Therefore the only
thing to do is to tell the app using libterm to reload the entire
screen.

The reason that updates to old_cur_screen are kept around even after
cur_screen has changed is because there is a chance that cur_screen could
be reset to what it was before the end of output processing. If there
was a way to make sure that this didn't happen, than update tracking and
storage could be completely turned off for the rest of the output
processing session. Howeverm there is no way to guarantee this, so
tracking and storage will continue. I suppose this is just a performance
vs. correctness issue.