Build system shuffle required to use libltdl (not actually used yet)
bootstrap.sh: Execute libtoolize with the --ltdl flag
configure.ac: Configure libltdl and set output variables
Makefile.am: Use LTDLDIR output variable to know whether to do recursive
		make into the libltdl directory
src/Makefile.am: Use LIBLTDL and LTDLINCL output variables to know how
		to link with libltdl and how to include the header

The output variables are different depending on whether a system or a
local copy of libltdl is being used.

Also note that, since the libtool I have is in the 1.5 series, some if
this usage is deprecated in the new 2.2 series of libtool. Will correct
this once Gentoo stabilizes the 2.2 series.