Only reference the data parameter of the error macros once.
The data parameter was being referenced twice (to check whether
it was not null and then to use if it it wasn't null) in the
error macro template macro. This causes problems when certain
function calls which return different things when called twice,
for example the lt_dlerror function which clears its error after
being called once, are passed as the data parameter.

To fix this, move the null check into its own function. This should
really be inlined, and I'm assuming that decent compilers will
automatically inline it. Also, I'm now defining a function in a
header file, which I've tried not to do in the past. This seems
right now to be the easiest cleanish fix, though.