Process the config file in a more detailed fashion, assuring its validity.
This causes the config file parsing code to parse each individual part
of the config key name, i.e. the module class name, the optional module
name, the name of the actual config entry, and the periods that separate
them. Previously, the whole key name was read in one single action.

This allows us to assure that the class name is valid and that there are
enough "parts" (i.e. things separated by periods) to the key.

Also, now that it no longer makes sense to have a function which sets
a config entry based on its key and new value, the set_config_entry
function was changed to take a class number, module name, and config
entry name. Since this function is basically the same as
ltm_set_config_entry without the init check and locking stuff, most of
ltm_set_config_entry it was deleted and now just calls set_config_entry.