Modularize the configuration subsystem
Add a new module class accordingly - MOD_CONF.

Since configuration is currently only done using flat text files,
this is currently the only module in existence. It should be fairly
easy to add new ones now, though.

By default, this module is the one loaded on initialization, unless
a different default was chosen with the --with-default-conf-module
flag to configure. In the future it will be possible to load a different
module at runtime, but this is currently not implemented.

This commit adds two header files for use by modules - libterm_module.h,
which is for all modules, and libterm_conf_module.h, which is for conf
modules. The module only includes the specific header, because the
general one will automatically be brought in at the same time and is not
designed for direct use. As new module classes are added, more headers
will be added.

The code for this module, and all modules in the future, is in
src/modules.

Add some checks to the config functions which explicitly disallow use of
the config system before it has been initialized.