Add a "cur_mid" global, for use in error reporting and configuration
This global stores the mid of the module which is currently being used.
The mechanism behind this is simply to set cur_mid before a call into
a module, and to set it back to -1 (which means "not in a module") when
the call is complete.

This value is used for error reporting and configuration. In error
reporting, it is used to report which module the error occurred in.
In configuration, it is used so that modules can call the functions
config_{get,set}_module_entry, which use the value of cur_mid instead
of taking the mid as an argument. This way, modules which are accessing
their own configurations don't have to keep track of their IDs.