Add another callback, "clear_screen"
This callback is, obviously, for instructing the program using
libterm to completely wipe the screen. It is currently used when
scrolling has caused all of the previous contents to no longer be
on the screen. In this case, the clear_screen callback is called
and then the accumulated updates are sent using the update_ranges
callback. Previously, the refresh_screen callback was used in this
case. However, I think it makes more sense to tell the program
using libterm specifically what changed.