Fix splitting up updates over the copy region
This didn't work at all before, because the condition was wrong.
If the condition had been correct, it would have mostly worked,
but still would have had some problems. The second part of the
split would have been too large - it would have extended to the
left and right edge of the screen, even though it only had to
extent to the left and right edges of the update. In addition,
the first part of the split would have been slightly too small,
not including enough of its last line.