Create an ACT_CLEAR update instead when everything has been scrolled off
I've had this for a while, but it was in a different place - in
the part of ltm_process_output which calls the callbacks. I only
recently realized that it really belonged here.

Note: an ignore-whitespace-changes diff for this commit will
more clearly show what changed than a normal diff.