Simply pass all updates to the program, even when everything would've been scrolled off
Before, I was trying to guess which of the following two things
would be faster, based on whether the region has been scrolled so
much that everything that was previously there has been scrolled
off:

1. Telling the program using libterm that the region was scrolled
   and then sending other updates to that region,
2. Telling the program using libterm to re-read the whole region.

With this change, always choose the former.

Note: an ignore-whitespace-changes diff for this commit will
more clearly show what changed than a normal diff.