Eliminate textparse.c's dependency on libterm
This involves not using libterm's typedefs and not
using libterm's error system. The latter is a bit tricky.
What I did was basically push the libterm error stuff up
the call stack, into the functions which call the textparse
functions. We use SYS_ERR to generate the error, because
this is a case where the function being called generates
an errno error, rather than a case where libterm itself
calculated that an incorrect case has been encountered
(which is what LTM_ERR is for). Even though these functions
aren't really "system" functions, it still fits better.