Merge the 'win_ups' and 'set' members of struct term_desc
The reason these were separate before was that I wanted
to be able to access the first propagated update from each
screen, because it specified the scroll region. I wanted
to be able to decide whether to pass the updates onto the
program using libterm or to just tell it to update the whole
region, based on whether the whole previous contents had
scrolled off. However, this has not been the behavior since
59717ffeb585c128bb3afaee240413aa545d7158. Therefore, there's
no point in keeping them separate anymore. Yay simplicity!