Store screen update information in struct screen
Before, update information for all screens was stored
directly in struct term_desc, in the scr_ups field.
However, this doesn't make much sense when one considers
that we already have a specific place for storing
information about screens. In addition, the previous method
required that we maintain the list of updates using
code in record_update. In particular, the whole list of
updated would have to be searched for the correct one
on every invocation. Now, we simply directly access the
updates for a particular screen.

The only downside to this new method that I can see is
that we now have to iterate over all of the screens in
order to find update information. However, this is only
done once per ltm_process_output call, and it should be
fast to skip over screens which have had no updates.

A lot of this patch has to do with reworking stuff to
take into account that the "sid" field of struct update
was removed. Since struct update is now only used in
struct screen, the sid is now implicit.