Add another point merge constraint
Say we have this screen, where the
included text has just been added:

+----------------------------------+
|The quick brown fox jumps over the|
|lazy                              |
|dog.                              |
+----------------------------------+

The precise string which was added
(in C syntax) was:

"The quick brown fox jumps over thelazy\ndog."

Before, the "thelazy\ndog" part would be
included in a single range. However, all of the
whitespace on the line with "lazy" would also
be included. This was happening because, when
choosing whether to merge a point on a new line
(such as the 'd' in dog) with the most recently
created range (such as the one with 'thelazy'),
we weren't checking to see whether the last line
on the most recent range extended all the way
out to its right boundary. In the 'thelazy' case,
the right boundary would be all the way to the
right edge of the screen, because of the line
wrap which happened after 'the'.

To fix this, only merge if the most recent
range extends all the way to its right
boundary, to avoid including unnecessary
whitespace.