Store update information for old_cur_screen in struct screen as well
Before, we needed a special case in record_update for when the original
window screen was being updated. With this change, we just store that
information in the window screen's struct update in its struct screen.
We can now get rid of the fields in struct term_desc and the special
handling code in record_update.

Most of this commit is just changing things from descs[tid].field to
CUR_SCR(tid).up.field or OLD_CUR_SCR(tid).up.field.