Get rid of record_update
record_update is now sufficiently simple that there's
really no point in keeping it around. Therefore, with this
commit we remove it and add the appropriate code in place
of any calls to it.

The only piece of code which would have to be copy-pasted
around is the condition of when to actually add the update.
Therefore, this condition has been added to a new SHOULD_UPDATE
macro.