Don't send SIGWINCH when window size changes.
I'm not sure why I didn't realize this before, but it seems that
sending SIGWINCH on window size change is actually done by the
kernel, in response to the TIOCSWINSZ ioctl.

This was tested by disabling and enabling the SIGWINCH sending
code in libterm. In both cases, the signal was still received
by the current foreground process group. In addition, the kernel
code seems to suggest that the kernel sends the signal.

I was initially planning on testing this by updating my kernel
with some extra tracing features, but that hasn't happened yet,
and the above test seems satisfactory to me.