Avoid a race condition involving SIGCHLD and when libterm figures out the PID of its children
If the shell process exits before libterm figures out its pid, the
SIGCHLD signal handler will be called but it will have no way of knowing
which terminal the signal applies to.

Solve this by blocking the SIGCHLD signal before calling spawn, and then
unblocking it once we've determined the PID and put it in a place where
the signal handler can find it.